Alex pointed out that we were not linking against the right things when -mabi=sandbox was still being passed to Clang.
This was because, while the compiler was outputting a warning and renaming
sandbox to purecap internally, the compiler driver was only expecting purecap
(when it should actually be looking out for both spellings). As a result, it
wasn't constructing the correct linker arguments.

This has now been fixed and a test has been added.