Output a warning in the case of MIPS CHERI if the callee doesn't have a prototype and we are passing arguments.
This would normally lead to using the variadic calling convention, which in the
case of MIPS CHERI, could lead to runtime stack corruption if the callee
function is not actually variadic.

The warning is enabled with the flag -Wmips-cheri-prototypes. It is enabled by
default when the target ABI is purecap. It can be turned into an error by passing
-Werror=mips-cheri-prototypes.